Revert "Pb 20 admin stock (#41)"
This reverts commit 42c0cdb880e544c6104d61a6b5315cdb9df92ed5.
