+++ front-end/src/App.js
@@ -9,7 +9,7 @@ import { AdminPage } from "./pages/AdminPage";
 import { VerificationPage } from "./pages/verificationPage";
 import DetailPage from "./pages/DetailPage";
 import { AdminDashboard } from "./pages/AdminDashboard";
-import cartSync from "./redux/cartSlice";
+import { cartSync } from "./redux/cartSlice";
 import { loanSync } from "./redux/loanSlice";
 import CartPage from "./pages/CartPage";
 import LoanPage from "./pages/LoanPage";
@@ -20,9 +20,6 @@ function App() {
   const token = localStorage.getItem("token");
   const tokenAdmin = localStorage.getItem("tokenAdmin");
 
-  const { NIM } = useSelector((state) => state.userSlice.value);
-
-
   const keepLogin = async () => {
     try {
       const res = await Axios.get(`http://localhost:2000/user/keepLogin`, {
@@ -31,17 +28,10 @@ function App() {
         },
       });
 
-
-      const result = await Axios.get(
-        `http://localhost:2000/cart/${res.data.NIM}`
-      );
-      dispatch(cartSync(result.data));
-
       const cart = await Axios.get(
         `http://localhost:2000/cart/${res.data.NIM}`
       );
       dispatch(cartSync(cart.data));
-
 
       const loan = await Axios.get(
         `http://localhost:2000/loan/${res.data.NIM}`
@@ -54,7 +44,7 @@ function App() {
           username: res.data.username,
           email: res.data.email,
           isVerified: res.data.isVerified,
-          cart: result.data.length,
+          cart: cart.data.length,
           loan: loan.data.length,
         })
       );
@@ -81,13 +71,11 @@ function App() {
   };
 
   useEffect(() => {
-
-    tokenAdmin
+        tokenAdmin
       ? keepLoginAdmin()
-      : token
+        : token
       ? keepLogin()
-      : console.log("Open Library");
-
+        : console.log("Open Library");
   });
 
   return (
