+++ front-end/src/components/AdminComp.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const AdminComp = () => {
+  return (
+    <div>AdminComp</div>
+  )
+}
